requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
alembic==1.13.2
asyncio==3.4.3
python-jose
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.32